backend/requirements.txt
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.3
fsspec==2026.1.0
gradio_client==2.0.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.3
idna==3.11
packaging==26.0
PyMySQL==1.1.2
python-dotenv==1.2.1
PyYAML==6.0.3
shellingham==1.5.4
sqlparse==0.5.5
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
